Refactor: Se cambio el modelo de Character porque el parámetro id viene desde la api como Long, no Int.

+++ app/src/main/java/dev/leotoloza/avengersapp/domain/model/Event.kt
@@ -1,9 +1,9 @@
 package dev.leotoloza.avengersapp.domain.model
 
 data class Event (
-    val id : Int,
-    val title : String,
-    val description : String,
+    val id: Long,
+    val title: String,
+    val description: String,
     val thumbnailUrl: String,
     val comics: List<Comic>,
 )
