+++ app/src/main/java/dev/leotoloza/avengersapp/ui/viewmodels/CharactersViewModel.kt
@@ -47,7 +47,7 @@ class CharactersViewModel
         }
     }
 
-    fun getCharacterById(id: Int): Character? {
+    fun getCharacterById(id: Long): Character? {
         return cachedCharacters.find { it.id == id }
     }
